requirements.txt
asgiref==3.10.0
certifi==2025.7.9
charset-normalizer==3.4.2
dig==0.1.10
dj-database-url==3.0.1
Django==5.2.7
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
docopt==0.6.2
futures==3.0.5
gunicorn==23.0.0
idna==3.10
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.7.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
